Conflicts:
	chainer_setup_build.py

+++ tests/cupy_tests/random_tests/test_generator.py
@@ -219,8 +219,10 @@ class TestInterval(unittest.TestCase):
     @condition.repeat(10)
     def test_within_interval(self):
         val = self.rs.interval(10, (2, 3)).get()
-        numpy.testing.assert_array_less(numpy.full((2, 3), -1), val)
-        numpy.testing.assert_array_less(val, numpy.full((2, 3), 11))
+        numpy.testing.assert_array_less(
+            numpy.full((2, 3), -1, dtype=numpy.int64), val)
+        numpy.testing.assert_array_less(
+            val, numpy.full((2, 3), 11, dtype=numpy.int64))
 
     @condition.retry(20)
     def test_lower_bound(self):
